feat: show sent messages in dashboard
